Nachal delat huinu; nedodelal

+++ src/someFunc.hpp
@@ -43,7 +43,7 @@ vector<double> getSpline(vector<double> n, int st, int fn) {
 
 double getDiff(graph a, graph b) {
 	double diff = 0;
-	for (int i = 0; i < a.x.size(); i++) {
+	for (size_t i = 0; i < a.x.size(); i++) {
 		diff = (diff > abs(a.y[i] - b.y[i])) ? diff : abs(a.y[i] - b.y[i]);
 	}
 	return diff;
